+++ jest.config.js
@@ -1,6 +1,10 @@
 module.exports = {
 	testEnvironment: 'node',
-	testPathIgnorePatterns: [ '/fixtures/' ],
+	testPathIgnorePatterns: [
+		'/fixtures/',
+		'/larva/src/patterns/',
+		'/larva-patterns/'
+	],
 	setupFiles: [
 		'./jest-setup.js'
 	]