+++ packages/larva-js/src/interface/MobileHeightToggle/MobileHeightToggle.js
@@ -1,23 +1,24 @@
 /**
  * Add drop-down functionality for navigation component.
  */
+
 export default class MobileHeightToggle {
 	constructor( el ) {
 		this.el = el;
 		this.el.pmcMobileHeightToggle = this;
 		this.isExpanded = false;
-		
+
 		this.toggle = this.toggle.bind( this );
 		this.onClick = this.onClick.bind( this );
 		this.onFocus = this.onFocus.bind( this );
 		this.onBlur = this.onBlur.bind( this );
 		this.keyDown = this.keyDown.bind( this );
-		
+
 		this.el.addEventListener( 'click', this.onClick );
 		this.el.addEventListener( 'focus', this.onFocus );
 		this.el.addEventListener( 'blur', this.onBlur );
 	}
-	
+
 	destroy() {
 		this.el.classList.remove( 'is-expanded' );
 		this.isExpanded = false;
@@ -28,7 +29,7 @@ export default class MobileHeightToggle {
 		document.removeEventListener( 'keydown', this.keyDown );
 	}
 
-	keyDown( e ) {	
+	keyDown( e ) {
 		if ( 'Enter' === e.key ) {
 			this.toggle( e );
 		}
@@ -39,7 +40,7 @@ export default class MobileHeightToggle {
 	}
 
 	onClick( e ) {
-		
+
 		if ( undefined !== this.el.pmcMobileHeightToggle ) {
 
 			// Bubble event