+++ packages/larva-patterns/components/c-icon/c-icon.prototype.js
@@ -1,6 +1,7 @@
 module.exports = {
 	c_icon_link_classes: '',
 	c_icon_name: 'byline-plus',
+	c_icon_target_attr: '_blank',
 	c_icon_link_screen_reader_text: 'Icon Link',
 	c_icon_url: false,
 	c_icon_classes: 'lrv-u-display-block lrv-u-width-16 lrv-u-height-16',