+++ packages/larva-patterns/modules/list/list.ordered.js
@@ -1,0 +1,7 @@
+const clonedeep = require( 'lodash.clonedeep' );
+
+const list = clonedeep( require( './list.prototype' ) );
+
+list.list_type = 'o';
+
+module.exports = list;